service/requirements.txt
Flask==2.3.3
google-cloud-firestore
gunicorn==21.2.0
